basic exec module
