+++ Makefile
@@ -25,6 +25,7 @@ SRC				= $(SRC_DIR)/main.c \
 					$(SRC_DIR)/ast/destructor.c \
 					$(SRC_DIR)/ast/getter.c \
 					$(SRC_DIR)/ast/push.c \
+					$(SRC_DIR)/exec/exec.c \
 					$(SRC_DIR)/lexer/constructor.c \
 					$(SRC_DIR)/lexer/get_next_token.c \
 					$(SRC_DIR)/lexer/is_start_of_input.c \
